Add warehouse-import components. Add filter for warehouse-product table.

+++ WarehouseWebMVC/Models/DTOs/ImportProductsDTO.cs
@@ -7,7 +7,7 @@ namespace WarehouseWebMVC.Models.DTOs
         public ICollection<Warehouse> ImportProducts { get; set; } = new List<Warehouse>();
         public long UserId { get; set; }
         public long SupplierId { get; set; }
-        public long Total { get; set; }
+        public double Total { get; set; }
         public string Reason { get; set; } = string.Empty;
         public string ReasonDetail { get; set; } = string.Empty;
         public string Deliverer { get; set; } = string.Empty;
